requirements.txt
# Core ML and Data Processing
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# NLP and Text Processing
transformers>=4.35.0
torch>=2.1.0
nltk>=3.8.1
textblob>=0.17.1
sentence-transformers>=2.2.2

# Database
pymongo>=4.5.0
motor>=3.3.1

# Web Dashboard
streamlit>=1.28.0
plotly>=5.16.1
altair>=5.1.2

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.1
python-dotenv>=1.0.0
click>=8.1.7

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0

# Development
black>=23.9.1
flake8>=6.1.0 